Significantly simplified program by using a different class for the Gui and the Board.
The Gui class handles all graphical things. i.e. Making buttons, setting the
colors of the buttons, etc.

The Board class is like the backend of the application. It creates the
board using a 2d array and has all the methods to check if there is a
win, etc.

I also simplified a lot of the code for checking if there is a win, creating
the buttons, etc.